fix : outlierValue를 int 자료형에서 double 자료형으로 변경

+++ src/main/java/live/ioteatime/apiservice/domain/Outlier.java
@@ -21,7 +21,7 @@ public class Outlier {
     @Column
     String type;
     @Column(name = "outlier_value")
-    int outlierValue;
+    double outlierValue;
     @Column
     int flag;
 }
